pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mdlint-obsidian"
version = "0.1.0"
description = "A linter for Obsidian Flavored Markdown files"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
keywords = ["obsidian", "markdown", "linter", "lint"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.0",
]

[project.scripts]
mdlint = "mdlint_obsidian.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mdlint_obsidian*"]
